+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/client/model/RetexturedModel.java
@@ -1,74 +1,99 @@
 package net.mehvahdjukaar.moonlight.api.client.model;
 
-import it.unimi.dsi.fastutil.objects.Object2ObjectArrayMap;
 import it.unimi.dsi.fastutil.objects.Object2ObjectLinkedOpenHashMap;
 import net.mehvahdjukaar.moonlight.api.client.util.VertexUtil;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.block.BlockModelShaper;
 import net.minecraft.client.renderer.block.model.BakedQuad;
 import net.minecraft.client.renderer.block.model.ItemOverrides;
 import net.minecraft.client.renderer.block.model.ItemTransforms;
-import net.minecraft.client.renderer.block.model.multipart.MultiPart;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.resources.model.BakedModel;
+import net.minecraft.client.resources.model.ModelResourceLocation;
 import net.minecraft.core.Direction;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 
-import java.util.IdentityHashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+import java.util.function.Function;
 
 public class RetexturedModel implements CustomBakedModel {
 
-    private static final Map<Block, RetexturedModel> parentCache = new Object2ObjectLinkedOpenHashMap<>();
+
+    private static final Map<ModelResourceLocation, RetexturedModel> parentCache = new Object2ObjectLinkedOpenHashMap<>();
 
     //TODO: reload this
-    private static final Map<BlockState, BakedModel> modelCache = new IdentityHashMap<>();
+    private static final Map<BlockState, Map<Direction, List<BakedQuad>>> modelCache = new IdentityHashMap<>();
 
-    public static RetexturedModel getInstance(Block block){
-        return parentCache.computeIfAbsent(block, RetexturedModel::new);
-    }
 
     private final Block parent;
+    private final boolean ambientOcclusion;
+    private Function<String, TextureAtlasSprite> newSpriteGetter;
+    private Map<Direction, List<String>> spriteOrder;
+    private final TextureAtlasSprite particle;
 
-    private RetexturedModel(Block parent) {
+    RetexturedModel(Block parent, Function<String, TextureAtlasSprite> newSpriteGetter,
+                    Map<Direction, List<String>> spriteOrder, boolean ambientOcclusion) {
         this.parent = parent;
-        MultiPart
+        this.newSpriteGetter = newSpriteGetter;
+        this.spriteOrder = spriteOrder;
+        this.particle = newSpriteGetter.apply("#particle");
+        this.ambientOcclusion = ambientOcclusion;
     }
 
     @Override
     public List<BakedQuad> getBlockQuads(BlockState state, Direction side, RandomSource rand, RenderType renderType, ExtraModelData extraModelData) {
+        if(state == null)state = parent.defaultBlockState();
+        var map = modelCache.computeIfAbsent(state, s -> new HashMap<>());
 
-        var shaper = Minecraft.getInstance().getModelManager().getBlockModelShaper();
+        var v = map.get(side);
+        if (v == null) {
 
-
-        BlockState parentState = parent.withPropertiesOf(state);
-        var parentModel = shaper.getBlockModel(parentState);
-
-        try {
-            var supportQuads = parentModel.getQuads(parentState, side, rand);
-            if (!supportQuads.isEmpty()) {
-                if (sprite != null) {
-                    supportQuads = VertexUtil.swapSprite(supportQuads, sprite);
-                }
-                quads.addAll(supportQuads);
+            if(spriteOrder == null || !spriteOrder.containsKey(side)){
+                map.put(side, List.of());
+                return List.of();
             }
 
-        } catch (Exception ignored) {
-        }
+            var shaper = Minecraft.getInstance().getModelManager().getBlockModelShaper();
+
+            BlockState parentState = parent.withPropertiesOf(state);
+            var parentModel = shaper.getBlockModel(parentState);
+            List<BakedQuad> quads = new ArrayList<>();
+
+            try {
+
+                var originalQuads = parentModel.getQuads(parentState, side, rand);
+                if (!originalQuads.isEmpty()) {
+                    var list = spriteOrder.get(side);
+                    int i = 0;
+                    for (var q : originalQuads) {
+                        var textureName = list.get(i);
+
+                        quads.add(VertexUtil.swapSprite(q, newSpriteGetter.apply(textureName)));
+                        i++;
+                    }
+                }
+            } catch (Exception ignored) {
+            }
+            spriteOrder.remove(side);
+            if (map.size() == 7) newSpriteGetter = null;
+            if (spriteOrder.isEmpty()){
+                spriteOrder = null;
+            }
+            map.put(side, quads);
+            return quads;
+        } else return v;
     }
 
     @Override
     public TextureAtlasSprite getBlockParticle(ExtraModelData extraModelData) {
-        return null;
+        return particle;
     }
 
     @Override
     public boolean useAmbientOcclusion() {
-        return false;
+        return ambientOcclusion;
     }
 
     @Override
@@ -88,11 +113,12 @@ public class RetexturedModel implements CustomBakedModel {
 
     @Override
     public ItemTransforms getTransforms() {
-        return null;
+        return ItemTransforms.NO_TRANSFORMS;
     }
 
     @Override
     public ItemOverrides getOverrides() {
-        return null;
+        return ItemOverrides.EMPTY;
     }
+
 }