+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/client/model/RetexturedModelLoader.java
@@ -1,25 +1,23 @@
 package net.mehvahdjukaar.moonlight.api.client.model;
 
-import com.google.common.base.Joiner;
-import com.google.common.collect.Lists;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
-import com.mojang.datafixers.util.Either;
 import net.minecraft.client.renderer.block.model.BlockElement;
 import net.minecraft.client.renderer.block.model.BlockElementFace;
 import net.minecraft.client.renderer.block.model.BlockModel;
-import net.minecraft.client.renderer.texture.MissingTextureAtlasSprite;
-import net.minecraft.client.renderer.texture.TextureAtlas;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.resources.model.*;
 import net.minecraft.core.Direction;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.GsonHelper;
+import net.minecraft.world.level.block.Block;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
-import java.util.Optional;
+import java.util.Map;
 import java.util.function.Function;
 
 /**
@@ -27,57 +25,50 @@ import java.util.function.Function;
  */
 public class RetexturedModelLoader implements CustomModelLoader {
 
-    private final Function<BakedModel, CustomBakedModel> factory;
-    private final String path;
-
-    public RetexturedModelLoader(String modelPath, Function<BakedModel, CustomBakedModel> bakedModelFactory) {
-        this.factory = bakedModelFactory;
-        this.path = modelPath;
+    public RetexturedModelLoader() {
     }
 
     @Override
     public CustomGeometry deserialize(JsonObject json, JsonDeserializationContext context) throws JsonParseException {
-        String string = GsonHelper.getAsString(json, "copy_from");
-        return new Geometry(ResourceLocation.tryParse(string));
+        ResourceLocation parentLoc = ResourceLocation.tryParse(GsonHelper.getAsString(json, "parent_model"));
+        ResourceLocation parent = ResourceLocation.tryParse(GsonHelper.getAsString(json, "parent_block"));
+        var b = BuiltInRegistries.BLOCK.getOptional(parent);
+        if (b.isEmpty()) throw new JsonParseException("Could not find block with id " + parent);
+        return new Geometry(b.get(), parentLoc);
     }
-
 
     private static class Geometry implements CustomGeometry {
 
-        private final ResourceLocation modelLoc;
+        private final ResourceLocation parentLoc;
+        private final Block parentBlock;
 
-        private Geometry(ResourceLocation model) {
-            this.modelLoc = model;
+        private Geometry(Block parentBlock, ResourceLocation model) {
+            this.parentLoc = model;
+            this.parentBlock = parentBlock;
         }
 
         @Override
         public CustomBakedModel bake(ModelBaker modelBaker, Function<Material, TextureAtlasSprite> spriteGetter, ModelState transform, ResourceLocation location) {
-            UnbakedModel model = modelBaker.getModel(modelLoc);
-            model.resolveParents(modelBaker::getModel);
-            if(model instanceof BlockModel bm) {
-                getSpriteOrder(bm);
-            }
-
-            return new RetexturedModel()
-            return getCustomBakedModel(modelBaker, spriteGetter, transform, location, model, modelLoc);
-        }
-
-
-        private final List<String> spriteOrder = new ArrayList<>();
-        private final List<String> spriteOrderUnculled = new ArrayList<>();
-
-
-        public void getSpriteOrder(BlockModel model) {
-            for (BlockElement blockElement : model.getElements()) {
-                for (Direction direction : blockElement.faces.keySet()) {
-                    BlockElementFace blockElementFace = blockElement.faces.get(direction);
-                    if (blockElementFace.cullForDirection == null) {
-                        spriteOrderUnculled.add(blockElementFace.texture);
-                    }else{
-                        spriteOrder.add(blockElementFace.texture);
+            BlockModel myModel = (BlockModel) modelBaker.getModel(location); // cursed
+            UnbakedModel parentModel = modelBaker.getModel(parentLoc);
+            parentModel.resolveParents(modelBaker::getModel);
+            Map<Direction, List<String>> spriteOrder = new HashMap<>();
+            if (parentModel instanceof BlockModel bm) {
+                for (BlockElement blockElement : bm.getElements()) {
+                    for (Direction direction : blockElement.faces.keySet()) {
+                        BlockElementFace blockElementFace = blockElement.faces.get(direction);
+                        if (blockElementFace.cullForDirection == null) {
+                            spriteOrder.computeIfAbsent(null, d -> new ArrayList<>())
+                                    .add(blockElementFace.texture);
+                        } else {
+                            spriteOrder.computeIfAbsent(direction, d -> new ArrayList<>())
+                                    .add(blockElementFace.texture);
+                        }
                     }
                 }
             }
+            Function<String, TextureAtlasSprite> newSpriteResolver = s-> spriteGetter.apply(myModel.getMaterial(s));
+            return new RetexturedModel(parentBlock, newSpriteResolver, spriteOrder, myModel.hasAmbientOcclusion());
         }
     }