+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/set/wood/WoodTypeRegistry.java
@@ -11,10 +11,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.Material;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.Collection;
-import java.util.IdentityHashMap;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 
 public class WoodTypeRegistry extends BlockTypeRegistry<WoodType> {
 
@@ -57,6 +54,8 @@ public class WoodTypeRegistry extends BlockTypeRegistry<WoodType> {
         return OAK_TYPE;
     }
 
+    public static Set<String> IGNORED_MODS = new HashSet<>(Set.of("chipped", "securitycraft", "absentbydesign"));
+
     //returns if this block is the base plank block
     @Override
     public Optional<WoodType> detectTypeFromBlock(Block baseBlock, ResourceLocation baseRes) {
@@ -85,8 +84,7 @@ public class WoodTypeRegistry extends BlockTypeRegistry<WoodType> {
             name = path.substring("plank_".length());
         }
         String namespace = baseRes.getNamespace();
-        if (name != null && !namespace.equals("securitycraft") &&
-                !namespace.equals("absentbydesign")) {
+        if (name != null && !IGNORED_MODS.contains(namespace)) {
 
             BlockState state = baseBlock.defaultBlockState();
             //can't check if the block is a full one, so I do this. Adding some checks here