+++ forge/src/main/java/net/mehvahdjukaar/moonlight/api/platform/forge/ClientHelperImpl.java
@@ -13,6 +13,7 @@ import net.minecraft.client.renderer.block.model.BlockModel;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.client.resources.model.ModelManager;
+import net.minecraft.client.resources.model.UnbakedModel;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleOptions;
 import net.minecraft.core.particles.ParticleType;
@@ -237,6 +238,10 @@ public class ClientHelperImpl {
                     return null;
                 }
         );
+    }
+
+    public static UnbakedModel getUnbakedModel(ModelManager modelManager, ResourceLocation modelLocation) {
+        return modelManager.getModelBakery().getModel(modelLocation);
     }